feat: Add link blocking functionality
This commit adds the ability to block links without removing them from the database. It introduces a new property `IsBlocked` to the `Link` model, which indicates whether a link is enabled and can be used. The migration file `20240419112310_AddLinkBlocking.cs` is created to add the necessary column to the `links` table in the database. Additionally, the `UserService` now includes a method `DisableUserAsync()` that allows disabling a user by setting their `IsBlocked` property to true.